Move RFID functionality to accounts app
