+++ accounts/models.py
@@ -3,6 +3,8 @@ from django.db import models
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
+from django.core.validators import RegexValidator
+from django.apps import apps
 from datetime import timedelta
 
 
@@ -167,6 +169,49 @@ class UserProxy(User):
         verbose_name_plural = User._meta.verbose_name_plural
 
 
+class RFID(models.Model):
+    """RFID tag that may be assigned to one or more accounts."""
+
+    rfid = models.CharField(
+        max_length=8,
+        unique=True,
+        verbose_name="RFID",
+        validators=[
+            RegexValidator(
+                r"^[0-9A-Fa-f]{8}$",
+                message="RFID must be 8 hexadecimal digits",
+            )
+        ],
+    )
+    allowed = models.BooleanField(default=True)
+    added_on = models.DateTimeField(auto_now_add=True)
+    is_seed_data = models.BooleanField(default=False)
+
+    def save(self, *args, **kwargs):
+        if self.rfid:
+            self.rfid = self.rfid.upper()
+        super().save(*args, **kwargs)
+        if not self.allowed:
+            self.accounts.clear()
+
+    def __str__(self):  # pragma: no cover - simple representation
+        return self.rfid
+
+    @staticmethod
+    def get_account_by_rfid(value):
+        """Return the account associated with an RFID code if it exists."""
+        Account = apps.get_model("accounts", "Account")
+        return (
+            Account.objects.filter(
+                rfids__rfid=value.upper(), rfids__allowed=True
+            )
+            .first()
+        )
+
+    class Meta:
+        verbose_name = "RFID"
+        verbose_name_plural = "RFIDs"
+        db_table = "accounts_rfid"
 
 
 class Account(models.Model):
@@ -180,7 +225,7 @@ class Account(models.Model):
         blank=True,
     )
     rfids = models.ManyToManyField(
-        "rfid.RFID", blank=True, related_name="accounts"
+        "RFID", blank=True, related_name="accounts"
     )
     service_account = models.BooleanField(
         default=False,
