+++ accounts/templatetags/rfid_tags.py
