+++ config/settings.py
@@ -49,7 +49,6 @@ INSTALLED_APPS = [
     "channels",
     "nodes",
     "accounts",
-    "rfid",
     "ocpp",
     "references",
     "awg",
