+++ ocpp/tests.py
@@ -12,7 +12,7 @@ from config.asgi import application
 
 from .models import Transaction, Charger, Simulator, MeterReading
 from accounts.models import Account, Credit
-from rfid.models import RFID
+from accounts.models import RFID
 from . import store
 from django.db.models.deletion import ProtectedError
 import json
